Update date and time example for terra-text
