+++ packages/terra-core-docs/src/terra-dev-site/test/text/DateTest.test.jsx
@@ -1,9 +1,44 @@
 import React from 'react';
 import Text from 'terra-text';
 import moment from 'moment';
+import classNames from 'classnames';
+import classNamesBind from 'classnames/bind';
+import styles from '../../../../../terra-text/src/Text.module.scss';
 
-const dateVar = moment("01-10-2022").format("MM/DD/YYYY");
+const cx = classNamesBind.bind(styles);
+const descriptiveClasses = classNames(
+  cx([
+    'text',
+    { 'visually-hidden': true },
+  ]),
+);
 
-export default () => <Text id="dateTest">
-    {dateVar}
-</Text>;
+const dateVar = moment('01-10-2022').format('MM/DD/YYYY');
+
+const isValidDate = moment(dateVar, 'MM/DD/YYYY', true).isValid();
+
+const dateParts = isValidDate ? dateVar.split('/') : [null, null, null];
+const monthVal = dateParts[0];
+const dayVal = dateParts[1];
+const yearVal = dateParts[2];
+
+const valueToDisplay = (
+  <>
+    <span className={descriptiveClasses}>Month</span>
+    <React.Fragment>
+      {monthVal}
+      /
+    </React.Fragment>
+
+    <span className={descriptiveClasses}>Day</span>
+    <React.Fragment>
+      {dayVal}
+      /
+    </React.Fragment>
+
+    <span className={descriptiveClasses}>Year</span>
+    <React.Fragment>{yearVal}</React.Fragment>
+  </>
+);
+
+export default () => <Text id="dateTest">{valueToDisplay}</Text>;
